Replaced KryoSerializer with an updated SparkSqlSerializer
* SparkSqlSerializer is moved to a separate source file
* SparkSqlSerializer.newKryo calls super.newKryo
* Class registration is no longer required, since we may de/serialise objects of any class with generic column accessor/builder.
